Se movio useRuthConeccedStore a RuthStore RuthStore sabe que es lo que envia Debvuelve promesas de evento
